Web Service System/Sever/web_service_system/requirement.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-jsonapi==6.1.0
inflection==0.5.1
Markdown==3.6
PyMySQL==1.1.0
sqlparse==0.5.0
